refactor: isolate sidebar rendering
